# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Spark Dating App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type LoginCredentialsLoginType = typeof LoginCredentialsLoginType[keyof typeof LoginCredentialsLoginType];


export const LoginCredentialsLoginType = {
  gmail: 'gmail',
  instagram: 'instagram',
} as const;

export interface LoginCredentials {
  loginType: LoginCredentialsLoginType;
  usernameOrEmail: string;
  password: string;
}

export type UserGender = typeof UserGender[keyof typeof UserGender];


export const UserGender = {
  Male: 'Male',
  Female: 'Female',
} as const;

export type UserInterestedIn = typeof UserInterestedIn[keyof typeof UserInterestedIn];


export const UserInterestedIn = {
  Male: 'Male',
  Female: 'Female',
} as const;

export type UserLoginType = typeof UserLoginType[keyof typeof UserLoginType];


export const UserLoginType = {
  gmail: 'gmail',
  instagram: 'instagram',
} as const;

export interface User {
  id: number;
  fullName: string;
  gender: UserGender;
  interestedIn: UserInterestedIn;
  age: number;
  city: string;
  bio: string;
  profileImage: string;
  loginType: UserLoginType;
  usernameOrEmail: string;
  interests: string[];
  verified: boolean;
  isOnline: boolean;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface UserUpdate {
  bio?: string;
  city?: string;
  interests?: string[];
}

export interface UserStats {
  likesSent: number;
  likesReceived: number;
  matches: number;
  superLikes: number;
  profileViews: number;
}

export type SwipeActionAction = typeof SwipeActionAction[keyof typeof SwipeActionAction];


export const SwipeActionAction = {
  like: 'like',
  pass: 'pass',
  superlike: 'superlike',
} as const;

export interface SwipeAction {
  targetUserId: number;
  action: SwipeActionAction;
}

export type SwipeResultAction = typeof SwipeResultAction[keyof typeof SwipeResultAction];


export const SwipeResultAction = {
  like: 'like',
  pass: 'pass',
  superlike: 'superlike',
} as const;

export interface SwipeResult {
  isMatch: boolean;
  action: SwipeResultAction;
  matchedUser?: User;
}

export interface Match {
  id: number;
  matchedAt: string;
  user: User;
}

export type DemoCredentialLoginType = typeof DemoCredentialLoginType[keyof typeof DemoCredentialLoginType];


export const DemoCredentialLoginType = {
  gmail: 'gmail',
  instagram: 'instagram',
} as const;

export interface DemoCredential {
  loginType: DemoCredentialLoginType;
  usernameOrEmail: string;
  password: string;
  fullName: string;
  profileImage: string;
}

export interface DemoCredentialsList {
  gmail: DemoCredential[];
  instagram: DemoCredential[];
}

export type DiscoverUsersParams = {
minAge?: number;
maxAge?: number;
city?: string;
};