# lib/api-zod/src/generated/types/demoCredential.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Spark Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { DemoCredentialLoginType } from './demoCredentialLoginType';

export interface DemoCredential {
  loginType: DemoCredentialLoginType;
  usernameOrEmail: string;
  password: string;
  fullName: string;
  profileImage: string;
}