# lib/api-zod/src/generated/types/demoCredentialsList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Spark Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { DemoCredential } from './demoCredential';

export interface DemoCredentialsList {
  gmail: DemoCredential[];
  instagram: DemoCredential[];
}