# lib/api-zod/src/generated/types/loginCredentials.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Spark Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginCredentialsLoginType } from './loginCredentialsLoginType';

export interface LoginCredentials {
  loginType: LoginCredentialsLoginType;
  usernameOrEmail: string;
  password: string;
}