# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Spark Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { UserGender } from './userGender';
import type { UserInterestedIn } from './userInterestedIn';
import type { UserLoginType } from './userLoginType';

export interface User {
  id: number;
  fullName: string;
  gender: UserGender;
  interestedIn: UserInterestedIn;
  age: number;
  city: string;
  bio: string;
  profileImage: string;
  loginType: UserLoginType;
  usernameOrEmail: string;
  interests: string[];
  verified: boolean;
  isOnline: boolean;
}